Rename app from canoe to src
