+++ pyproject.toml
@@ -1,5 +1,5 @@
 [project]
-name = 'canoe'
+name = 'src'
 version = '2.1.3'
 description = 'The second release of the Sheikhs application'
 dependencies = [
