+++ src/auth/views.py
@@ -12,9 +12,9 @@ from werkzeug.exceptions import abort
 from werkzeug.security import check_password_hash
 from werkzeug.security import generate_password_hash
 
-from canoe.auth import bp
-from canoe.auth.models import Users
-from canoe.extensions import db
+from src.auth import bp
+from src.auth.models import Users
+from src.extensions import db
 
 
 @bp.route('/register', methods=('GET', 'POST'))
