+++ src/gallery/__init__.py
@@ -2,4 +2,4 @@ from flask import Blueprint
 
 bp = Blueprint('gallery', __name__, template_folder='templates')
 
-from canoe.gallery import views
+from src.gallery import views
