+++ src/gallery/templates/create.html
