+++ src/gallery/templates/detail.html
