+++ src/gallery/templates/index.html
