+++ src/gallery/templates/update.html
