+++ src/gallery/views.py
@@ -10,10 +10,10 @@ from flask import request
 from flask import url_for
 from werkzeug.exceptions import abort
 
-from canoe.auth.views import login_required
-from canoe.extensions import db
-from canoe.gallery import bp
-from canoe.gallery.models import Work
+from src.auth.views import login_required
+from src.extensions import db
+from src.gallery import bp
+from src.gallery.models import Work
 
 
 @bp.route('/')
