+++ src/wsgi.py
@@ -1,4 +1,4 @@
-from canoe import create_app
+from src import create_app
 
 
 if __name__ == '__main__':
