+++ tests/test_auth.py
@@ -2,7 +2,7 @@ from flask import g
 from flask import session
 
 from .conftest import DEFAULT_USER
-from canoe.auth.models import Users
+from src.auth.models import Users
 
 
 def test_register_get_request(app, client):
